% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinearbart.R
\name{loglinearbart}
\alias{loglinearbart}
\title{loglinearbart}
\usage{
loglinearbart(
  x.train,
  y.train,
  de.train,
  wi.train,
  v.train,
  x.test = matrix(0, 0, 0),
  power = 2,
  base = 0.95,
  ntree = 200,
  ndpost = 1000,
  nskip = 100,
  printevery = 100,
  keepevery = 1,
  keeptrainfits = TRUE,
  usequants = FALSE,
  numcut = 100,
  printcutoffs = 1,
  verbose = TRUE,
  aa = 4.5,
  bb = 4
)
}
\arguments{
\item{x.train}{Description of arg1}

\item{y.train}{Description of arg2}

\item{de.train}{Description of arg1}

\item{wi.train}{Description of arg2}

\item{v.train}{Description of arg1}

\item{x.test}{Description of arg2}

\item{power}{Description of arg2}

\item{base}{Description of arg1}

\item{ntree}{Description of arg1}

\item{ndpost}{Description of arg2}

\item{nskip}{Description of arg1}

\item{printevery}{Description of arg2}

\item{keepevery}{Description of arg1}

\item{keeptrainfits}{Description of arg2}

\item{usequants}{Description of arg1}

\item{numcut}{Description of arg2}

\item{printcutoffs}{Description of arg1}

\item{verbose}{Description of arg2}

\item{aa}{Description of arg2}

\item{bb}{Description of arg2
this is a function}
}
\description{
loglinearbart
}
